FIXED: Windows compiler warnings and build issues in podofoimpose
git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/trunk@1197 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5

+++ src/PdfXRefStreamParserObject.cpp
@@ -109,7 +109,7 @@ void PdfXRefStreamParserObject::ParseStream( const pdf_int64 nW[W_ARRAY_SIZE], c
 {
     char*        pBuffer;
     pdf_long     lBufferLen;
-    const size_t entryLen  = (nW[0] + nW[1] + nW[2]);
+    const size_t entryLen  = static_cast<size_t>(nW[0] + nW[1] + nW[2]);
 
     this->GetStream()->GetFilteredCopy( &pBuffer, &lBufferLen );
